Updates 12/3
Realized that the TemplateCheck macro was glitching because of MatchCase features.
On the one hand, I got it to successfully ignore when it finds "insert" in lowercase. On the other hand, I've broken the code for the first person check and I don't know how to get it back.
The punctuation macro wasn't working well with regards to , and "
Inserted extra search function to clear formatting.